Ignore pre_delete signal if model is not tracked

+++ avam/categories/signals.py
@@ -1,5 +1,6 @@
 from django.db.models.signals import post_save, pre_delete
 from django.dispatch import receiver
+from django.contrib.contenttypes.models import ContentType
 
 from categories.models import Category, CategoryItem
 
@@ -19,5 +20,9 @@ def on_all_pre_delete(sender, **kwargs):
     if sender is Category:
         return
     instance = kwargs.get('instance')
+    m = instance._meta.model
+    content_type = ContentType.objects.get_for_model(m)
+    if not CategoryItem.objects.filter(content_type=content_type).exists():
+        return
     q = CategoryItem.objects.get_for_object(instance)
     q.delete()
